server/requirements.txt
Django==2.2.1
django_filters==0.2.1
djangorestframework==3.9.4
django-cors-headers==3.0.2